Things

+++ My project/Assets/Scripts/Generic/Variables.cs
@@ -7,8 +7,8 @@ namespace FrogTime
     internal class CharacterVariables : MonoBehaviour
     {
         //Character Variables 
-        internal int characterMaxHealt = 20;
-       internal int characterHealt = 20;
+       internal float characterMaxHealt = 20f;
+       internal float characterHealt = 20f;
        internal float characterSpeed;
        internal int characterDamage = 2;
        internal static float characterMoveSpeed = 0.1f;
